+++ example.py
@@ -9,13 +9,17 @@ def load_data(data):
     # Load the data
     with open('example.csv') as f:
         reader = csv.DictReader(f, fieldnames=['name', 'salary'])
-        reader.next()
+        next(reader)
         data['table'] = list(reader)
+
+
+def filter_rows(row):
+    return int(row['salary']) < 25000
 
 
 def select_rows(data):
     # Select relevant rows from the table
-    data['low_income'] = filter(lambda r: int(r['salary']) < 20000, data['table'])
+    data['low_income'] = list(filter(filter_rows, data['table']))
 
 
 def calculate_average(data):